- Add Form validation with Hibernate - Change Integers to BigIntegers, Unit tests updated

+++ oh-my-gacha/src/test/java/com/ohmygacha/OhMyGachaApplicationTests.java
@@ -3,6 +3,7 @@ package com.ohmygacha;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -27,9 +28,9 @@ class OhMyGachaApplicationTests {
 	@Test
 	void totalCostTest() {
 		var calc = new Calculator();
-		assertEquals(1000.0, calc.totalCost(200, new BigDecimal("99.99"), 2000, 100), "200 attempts should cost £1000.00");
-		assertEquals(750.0, calc.totalCost(500, new BigDecimal("38.99"), 13000, 500), "500 attempts should cost £750.00");
-		assertEquals(23.0, calc.totalCost(20, new BigDecimal("8.99"), 2000, 250), "20 attempts should cost £23.00");
+		assertEquals(BigInteger.valueOf(1000), calc.totalCost(200, new BigDecimal("99.99"), BigInteger.valueOf(2000), BigInteger.valueOf(100)), "200 attempts should cost £1000.00");
+		assertEquals(BigInteger.valueOf(750), calc.totalCost(500, new BigDecimal("38.99"), BigInteger.valueOf(13000), BigInteger.valueOf(500)), "500 attempts should cost £750.00");
+		assertEquals(BigInteger.valueOf(23), calc.totalCost(20, new BigDecimal("8.99"), BigInteger.valueOf(2000), BigInteger.valueOf(250)), "20 attempts should cost £23.00");
 		
 	}
